requirements.txt
pyyaml~=6.0.1
attrs~=23.2.0
pytest
trio~=0.25.0
trio-parallel
trio-typing
trio-websocket

setuptools~=69.2.0
Cython~=3.0.10
paho-mqtt~=2.0.0